Remove Ansible MrP dep, add MrP client
This role required ansible-role-mr-provisioner to be installed in the
filesystem without any mention of it. Not only that, but now we have the
client which is what that ansible role calls, so removing one
indirection, and making sure we checkout the repo instead of needing it
to be silently installed first.